Fix: Compile error on Linux

+++ src/Servers.cpp
@@ -1735,7 +1735,7 @@ void AdminServer::rpc_getinfo(Client *c, const RPC::Message &m)
     res["txs_sent"] = qulonglong(srvmgr->txBroadcasts());
     res["txs_sent_bytes"] = qulonglong(srvmgr->txBroadcastBytes());
     res["uptime"] = QString::number(Util::getTimeSecs(), 'f', 1) + " secs";
-    res["subscriptions"] = storage->subs()->numActiveClientSubscriptions();
+    res["subscriptions"] = qlonglong(storage->subs()->numActiveClientSubscriptions());
     res["peers"] = peers.size();
     res["config"] = options->toMap();
     { // mempool
